# web-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Loan Management System API",
        "description": "API documentation for Loan Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/payment/loans/{loan}/initiate": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Initiate payment for loan installment",
                "description": "Initialize payment for a loan installment",
                "operationId": "initiatePayment",
                "parameters": [
                    {
                        "name": "loan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "schedule_id"
                                ],
                                "properties": {
                                    "schedule_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": 1000000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment initiated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "authority": {
                                            "type": "string"
                                        },
                                        "payment_url": {
                                            "type": "string",
                                            "format": "url"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/payment/callback": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Payment gateway callback",
                "description": "Handle callback from payment gateway",
                "operationId": "paymentCallback",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Authority": {
                                        "type": "string"
                                    },
                                    "Status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment processed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/payment/status/{payment}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment status",
                "description": "Get payment status",
                "operationId": "getPaymentStatus",
                "parameters": [
                    {
                        "name": "payment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/payment/history": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment history",
                "description": "Get user's payment history",
                "operationId": "getPaymentHistory",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment history",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}